Changes made on Date : 2026-06-06

+++ Programming Languages/Clion C Family/C++/Projects/Server/Hash.hpp
@@ -1,20 +1,20 @@
 #include <string>
 
 struct Hash {
-    int findHash(std::string &s) {
-        int n = s.length();
+    long long findHash(std::string &s) {
+        long long n = s.length();
         // p is a prime number
         // m is a large prime number
-        int p = 31, m = 1e9 + 7;
+        long long p = 31, m = 1e9 + 7;
 
         // to store hash value
-        int hashVal = 0;
+        long long hashVal = 0;
 
         // to store p^i
-        int pPow = 1;
+        long long pPow = 1;
 
         // Calculating hash value
-        for (int i = 0; i < n; ++i) {
+        for (long long i = 0; i < n; ++i) {
             hashVal = (hashVal + (s[i] - 'a' + 1) * pPow) % m;
             pPow = (pPow * p) % m;
         }
